Added basic support for multiple maps.
Added/moved a few functions to the Map class.
Split FormulateSaveData function into two seperate GetSaveData functions
in the Level and Map class.
Fixed a placement bug introduced in the last commit where the placed
object was not a copy but a reference to the selected object.
Removed some redundant functions.